scheduler/requirements.txt
APScheduler>=3.10.0
PyYAML>=6.0
